Do the line transformation in one pass on file import.
Instead of creating two different Vec calling collect twice, this commit
do the processing in series before converting into a Vec.

This commit also removes the type suggestion for collect, as it is
capable of infering it now by the usage.

This is on top of PR #10 (Uses the `split_whitespace` iterator.)